Created CMWElecSectionFilter.cs class for repeated Selection Filters; updated PanelTypeToSinglePhase.cs to include DetailItemSelectionFilter.
